Query Sentinel imagery

user inputs a longitude,latitude and a width and height

scene(
  x = cbind(146.614867, -43.298699),
  date = Sys.Date() + c(-14, -2),
  wh = c(5000),
  proj = NULL,
  res = max(c(10, wh/1024)),
  silent = FALSE,
  dry_run = FALSE
)

Arguments

x

location

date

vector of date(-times)

wh

width height in metres around 'location'

proj

projection for scene

res

resolution desired (a default for a reasonable image size is provided)

silent

emit messages

dry_run

if TRUE only return the stac query

Value

data frame of available scenes and some processed links

Details

Scenes are returned in the decreasing order of file size of the processed local scene.

Examples

scene(dry_run = TRUE)
#> processing 3 scenes from 3 (solar) days
#> in longlat region lonmin,lonmax,latmin,latmax 146.584047,146.645687,-43.321201,-43.276192
#> # A tibble: 3 × 24
#>   red    green blue  visual nir   swir22 rededge2 rededge3 rededge1 swir16 wvp  
#>   <chr>  <chr> <chr> <chr>  <chr> <chr>  <chr>    <chr>    <chr>    <chr>  <chr>
#> 1 https… http… http… https… http… https… https:/… https:/… https:/… https… http…
#> 2 https… http… http… https… http… https… https:/… https:/… https:/… https… http…
#> 3 https… http… http… https… http… https… https:/… https:/… https:/… https… http…
#> # ℹ 13 more variables: nir08 <chr>, scl <chr>, aot <chr>, coastal <chr>,
#> #   nir09 <chr>, cloud <chr>, snow <chr>, preview <chr>,
#> #   granule_metadata <chr>, tileinfo_metadata <chr>, product_metadata <chr>,
#> #   thumbnail <chr>, solarday <date>